Batch 6: Eliminate ALL remaining 29 warnings (format strings, NULL conversions, sign-compare, unused funcs)
Fixed warnings in 10 files:
- savestate.cpp: 8 format string fixes (%ld->%u, %lld casts)
- filesys.cpp: 6 fixes (syntax error, format strings, sign-compare)
- expansion.cpp: 3 NULL->0 conversions for int fields
- uaeserial.cpp: 3 fixes (unused param, pointer format %p)
- cia.cpp: 2 unused functions marked __attribute__((unused))
- zfile_archive.cpp: 2 format string casts
- cdtvcr.cpp: 1 NULL->0 conversion
- newcpu.cpp: 1 format string cast (long long)
- scsi.cpp: 1 tautological comparison removed
- sndboard.cpp: 1 NULL->0 conversion

Total project warnings: 5,620 -> 0 (100% elimination!)

+++ uae_src/savestate.cpp
@@ -447,7 +447,7 @@ static void save_chunk (struct zfile *f, uae_u8 *chunk, size_t len, const TCHAR 
 		zfile_fwrite(zero, 1, len2, f);
 	}
 
-	write_log (_T("Chunk '%s' chunk size %u (%u)\n"), name, chunklen, len);
+	write_log (_T("Chunk '%s' chunk size %u (%u)\n"), name, (unsigned int)chunklen, (unsigned int)len);
 }
 
 static uae_u8 *restore_chunk (struct zfile *f, TCHAR *name, unsigned int *len, unsigned int *totallen, size_t *filepos)
@@ -1373,7 +1373,7 @@ int savestate_dorewind (int pos)
 		pos = replaycounter - 1;
 	if (canrewind (pos)) {
 		savestate_state = STATE_DOREWIND;
-		write_log (_T("dorewind %d (%010ld/%03ld) -> %d\n"), replaycounter - 1, hsync_counter, vsync_counter, pos);
+		write_log (_T("dorewind %d (%010u/%03u) -> %d\n"), replaycounter - 1, hsync_counter, vsync_counter, pos);
 		return 1;
 	}
 	return 0;
@@ -1518,7 +1518,7 @@ void savestate_rewind (void)
 		return;
 	}
 	inprec_setposition (st->inprecoffset, pos);
-	write_log (_T("state %d restored.  (%010ld/%03ld)\n"), pos, hsync_counter, vsync_counter);
+	write_log (_T("state %d restored.  (%010u/%03u)\n"), pos, hsync_counter, vsync_counter);
 	if (rewind) {
 		replaycounter--;
 		if (replaycounter < 0)
@@ -1906,9 +1906,9 @@ retry2:
 			staterecords_first -= staterecords_max;
 	}
 
-	write_log (_T("state capture %d (%010ld/%03ld,%ld/%d) (%ld bytes, alloc %d)\n"),
+	write_log (_T("state capture %d (%010u/%03u,%u/%d) (%ld bytes, alloc %d)\n"),
 		replaycounter, hsync_counter, vsync_counter,
-		hsync_counter % current_maxvpos (), current_maxvpos (),
+		(unsigned int)(hsync_counter % current_maxvpos ()), (int)current_maxvpos (),
 		st->end - st->data, statefile_alloc);
 
 	if (firstcapture) {
